+++ flake.nix
@@ -2,10 +2,8 @@
   description = "Nix shell for Holochain app development";
 
   inputs = {
+    holonix.url = "github:holochain/holonix/main";
     nixpkgs.follows = "holonix/nixpkgs";
-    versions.url = "github:holochain/holochain?dir=versions/weekly";
-    holonix.url = "github:holochain/holochain";
-    holonix.inputs.versions.follows = "versions";
   };
 
   outputs = inputs@{ holonix, ... }:
@@ -13,13 +11,15 @@
       # provide a dev shell for all systems that the holonix flake supports
       systems = builtins.attrNames holonix.devShells;
 
-      perSystem = { config, system, pkgs, ... }:
+      perSystem = { inputs', pkgs, ... }:
         {
           devShells.default = pkgs.mkShell {
-            inputsFrom = [ holonix.devShells.${system}.holochainBinaries ];
-            packages = with pkgs; [
+            packages = [
+              inputs'.holonix.packages.holochain
+              inputs'.holonix.packages.lair-keystore
+              inputs'.holonix.packages.rust
               # add further packages from nixpkgs
-              nodejs
+              pkgs.nodejs
             ];
           };
         };